Initial commit
The library was split out into its own repository from an NPM workspace in
tc39/proposals-array-from-async. This is to be able to maintain the library if
the proposal reaches Stage 4 (in which case the proposal repository would be
archived).

In addition, `package.json`’s metadata were fixed (version and license had not
been updated).

See <https://github.com/tc39/proposal-array-from-async/>
for preceding Git history.